+++ src/lib/components/Molecule3dViewer.js
@@ -85,9 +85,44 @@ Molecule3dViewer.propTypes = {
     selectedAtomIds: PropTypes.array,
 
     /**
-     * labels corresponding to the atoms of the molecule
+     * Labels corresponding to the atoms of the molecule.
+     * Each label has a `text` field, a string containing the label content,
+     * and can have many other styling fields as described in
+     * https://3dmol.csb.pitt.edu/doc/types.html#LabelSpec
      */
-    labels: PropTypes.array,
+    labels: PropTypes.arrayOf(PropTypes.object),
+
+    /**
+     * Add an isosurface from volumetric data provided in the `cube_file`
+     */
+    orbital: PropTypes.exact({
+        /**
+         * The filepath containing raw volumetric data for vertex coloring
+         */
+        cube_file: PropTypes.string,
+        /**
+         * The isovalue to draw the surface at
+         */
+        iso_val: PropTypes.number,
+        /**
+         * Transparency of the surface, between 0 and 1
+         */
+        opacity: PropTypes.number,
+        /**
+         * Color for the positive value of the isosurface orbital
+         */
+        positiveVolumetricColor: PropTypes.string,
+        /**
+         * Color for the negative value of the isosurface orbital
+         */
+        negativeVolumetricColor: PropTypes.string,
+    }),
+
+    /**
+     * Add a predefined renderable shape objects to the molecule.
+     * Valid shape types are Arrow, Sphere, and Cylinder.
+     */
+    shapes: PropTypes.arrayOf(PropTypes.object),
 
     /**
      * Callback to re-render molecule viewer