+++ tests/integration/test_ngl_moleculeviewer.py
@@ -15,7 +15,7 @@ import dash_bio
 
 # from dash_bio_utils import pdb_parser as parser
 
-from common_features import simple_app_layout, simple_app_callback
+from common_features import simple_app_layout, simple_app_callback, generate_identifier
 
 _COMPONENT_ID = "test-ngl"
 data_path = "tests/dashbio_demos/dash-ngl-moleculeviewer/data/"
@@ -264,7 +264,8 @@ def modified_simple_app_callback(
 
     time.sleep(5)
     if take_snapshot:
-        dash_duo.percy_snapshot(f"{component_id}_{test_prop_name}_{test_prop_value}")
+        dash_duo.percy_snapshot(f"{component_id}_{test_prop_name}_{test_prop_value}_"
+                                f"{generate_identifier()}", convert_canvases=True)
 
 
 def test_dbdn001_viewer_loaded(dash_duo):
@@ -568,7 +569,7 @@ def test_dbn_015_rotate_stage(dash_duo):
 
 def test_dbn_016_reset_stageView(dash_duo):
 
-    test_value = "6CHG.A_3K8P.D"
+    test_value = "6CHG.A_3K8P.C"
 
     app = dash.Dash(__name__)
 
@@ -587,7 +588,7 @@ def test_dbn_016_reset_stageView(dash_duo):
 
 def test_dbn_017_download_image(dash_duo):
 
-    test_value = "6CHG.A_3K8P.D"
+    test_value = "6CHG.A_3K8P"
 
     app = dash.Dash(__name__)