+++ tests/integration/test_speck.py
@@ -8,7 +8,7 @@ import dash_bio
 import dash_html_components as html
 from dash_bio_utils import xyz_reader
 
-from common_features import simple_app_layout, simple_app_callback
+from common_features import simple_app_layout, simple_app_callback, generate_identifier
 
 _data = None
 
@@ -26,7 +26,10 @@ def test_dbsp001_rotate(dash_duo):
     app.layout = html.Div(
         dash_bio.Speck(
             id=_COMPONENT_ID,
-            data=_data
+            data=_data,
+            view={
+                "brightness": 0
+            }
         )
     )
 
@@ -38,7 +41,7 @@ def test_dbsp001_rotate(dash_duo):
     ac.move_to_element(speck).drag_and_drop_by_offset(
         speck, 150, 200).perform()
 
-    dash_duo.percy_snapshot('test-speck_rotate')
+    dash_duo.percy_snapshot('test-speck_rotate' + generate_identifier(), convert_canvases=True)
 
 
 def test_dbsp002_click_and_drag(dash_duo):
@@ -60,7 +63,8 @@ def test_dbsp002_click_and_drag(dash_duo):
     ac.move_to_element(speck).key_down(Keys.SHIFT).drag_and_drop_by_offset(
         speck, -50, 100).key_up(Keys.SHIFT).perform()
 
-    dash_duo.percy_snapshot('test-speck_click_and_drag')
+    dash_duo.percy_snapshot('test-speck_click_and_drag' + generate_identifier(),
+                            convert_canvases=True)
 
 
 def test_dbsp003_preset_view_licorice(dash_duo):
@@ -148,7 +152,8 @@ def test_dbsp006_custom_view(dash_duo):
         'atomScale': 0.15,
         'relativeAtomScale': 0.51,
         'bonds': True,
-        'bondScale': 0.75
+        'bondScale': 0.75,
+        'brightness': 0
     }
 
     simple_app_callback(